add geolocation marker based on user location
